Removed test files, cleaned up console.logs and removed steps
Removed Bookmark and Jumbotron components. Removed console.logs and unused imports/varibales.

Updated Create Itinerary and Activity flows to go page to page rather than using steps. Removed reference of steps in App.js
