--- client/src/components/Bookmark/Bookmark.css
@@ -1,8 +1,0 @@
-:root{
-    /* Color Palette 3 */
-    --Lotus-Green: #317312;
-    --Green-Scallion: #6EBB2D;
-    --Wafer-Brown: #DED6C9;
-    --Blue-Isle: #24BAEC;
-    --Capri-Blue: #166EF3;
- }
