+++ client/src/components/CreateActivity/Create activity.png
