+++ client/src/components/Rating/index.js
@@ -1,123 +1,109 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import { withStyles } from '@material-ui/core/styles';
-import { UPDATE_ACTIVITY } from "../../utils/mutations"
-import { QUERY_ACTIVITY_BY_ID } from "../../utils/queries"
-import { useMutation } from '@apollo/react-hooks';
-import { useQuery } from '@apollo/react-hooks';
-import Rating from '@material-ui/lab/Rating';
-import StarBorderIcon from '@material-ui/icons/StarBorder';
-import SentimentVeryDissatisfiedIcon from '@material-ui/icons/SentimentVeryDissatisfied';
-import SentimentDissatisfiedIcon from '@material-ui/icons/SentimentDissatisfied';
-import SentimentSatisfiedIcon from '@material-ui/icons/SentimentSatisfied';
-import SentimentSatisfiedAltIcon from '@material-ui/icons/SentimentSatisfiedAltOutlined';
-import SentimentVerySatisfiedIcon from '@material-ui/icons/SentimentVerySatisfied';
-import Typography from '@material-ui/core/Typography';
-import Box from '@material-ui/core/Box';
+// import React from 'react';
+// import PropTypes from 'prop-types';
+// import { withStyles } from '@material-ui/core/styles';
+// import { UPDATE_ACTIVITY } from "../../utils/mutations"
+// import { QUERY_ACTIVITY_BY_ID } from "../../utils/queries"
+// import { useMutation } from '@apollo/react-hooks';
+// import { useQuery } from '@apollo/react-hooks';
+// import Rating from '@material-ui/lab/Rating';
+// import StarBorderIcon from '@material-ui/icons/StarBorder';
+// import SentimentVeryDissatisfiedIcon from '@material-ui/icons/SentimentVeryDissatisfied';
+// import SentimentDissatisfiedIcon from '@material-ui/icons/SentimentDissatisfied';
+// import SentimentSatisfiedIcon from '@material-ui/icons/SentimentSatisfied';
+// import SentimentSatisfiedAltIcon from '@material-ui/icons/SentimentSatisfiedAltOutlined';
+// import SentimentVerySatisfiedIcon from '@material-ui/icons/SentimentVerySatisfied';
+// import Typography from '@material-ui/core/Typography';
+// import Box from '@material-ui/core/Box';
 
 
-const customIcons = {
-  1: {
-    icon: <SentimentVeryDissatisfiedIcon />,
-    label: 'Very Dissatisfied',
-  },
-  2: {
-    icon: <SentimentDissatisfiedIcon />,
-    label: 'Dissatisfied',
-  },
-  3: {
-    icon: <SentimentSatisfiedIcon />,
-    label: 'Neutral',
-  },
-  4: {
-    icon: <SentimentSatisfiedAltIcon />,
-    label: 'Satisfied',
-  },
-  5: {
-    icon: <SentimentVerySatisfiedIcon />,
-    label: 'Very Satisfied',
-  },
-};
+// const customIcons = {
+//   1: {
+//     icon: <SentimentVeryDissatisfiedIcon />,
+//     label: 'Very Dissatisfied',
+//   },
+//   2: {
+//     icon: <SentimentDissatisfiedIcon />,
+//     label: 'Dissatisfied',
+//   },
+//   3: {
+//     icon: <SentimentSatisfiedIcon />,
+//     label: 'Neutral',
+//   },
+//   4: {
+//     icon: <SentimentSatisfiedAltIcon />,
+//     label: 'Satisfied',
+//   },
+//   5: {
+//     icon: <SentimentVerySatisfiedIcon />,
+//     label: 'Very Satisfied',
+//   },
+// };
 
-function IconContainer(props) {
-  const { value, ...other } = props;
-  return <span {...other}>{customIcons[value].icon}</span>;
-}
+// function IconContainer(props) {
+//   const { value, ...other } = props;
+//   return <span {...other}>{customIcons[value].icon}</span>;
+// }
 
-IconContainer.propTypes = {
-  value: PropTypes.number.isRequired,
-};
-
-
-const CustomizedRatings = (props) => {
-  console.log('hello', props.id)
-  // const [updateActivity, { error }] = useMutation(UP_ACTIVITY);
-
-  const [value, setValue] = React.useState(4);
-  const [hover, setHover] = React.useState(-1);
-  const labels = {
-    0.5: 'Useless',
-    1: 'Useless+',
-    1.5: 'Poor',
-    2: 'Poor+',
-    2.5: 'Ok',
-    3: 'Ok+',
-    3.5: 'Good',
-    4: 'Good+',
-    4.5: 'Excellent',
-    5: 'Excellent+',
-  }
-  // const = useState({ rating: $rating })
-  const [updateActivity, { error }] = useMutation(UPDATE_ACTIVITY);
-  const itineraryId = localStorage.getItem("itineraryId");
-  console.log(itineraryId)
-
-  const handleFormSubmit = async event => {
-    event.preventDefault();
-    try {
-      await updateActivity({
-        variables: {
-          rating: props.rating,
-        }
-      })
-      console.log(updateActivity)
-    }
-    catch (e) {
-      console.log(e);
-    }
-  };
+// IconContainer.propTypes = {
+//   value: PropTypes.number.isRequired,
+// };
 
 
-  return (
-    <div>
-      <form onSubmit={handleFormSubmit}>
+// const CustomizedRatings = (props) => {
+//   console.log('hello', props.id)
+//   // const [updateActivity, { error }] = useMutation(UP_ACTIVITY);
 
-        <Box component="fieldset" mb={3} borderColor="transparent">
-          <Typography component="legend">How do you like this activity?</Typography>
-          <Rating
-            name="simple-controlled"
-            value={value}
-            onChange={(event, newValue) => {
-              setValue(newValue);
-              console.log("newValue", newValue);
-              //update itinerary using newValue as a parameter, 
-              // updateActivity({
-              //   variables: {
-              //     rating: props.box
-              //   }
-              // })
-              // updateItinerary(newValue, props.id)
+//   const [value, setValue] = React.useState(4);
+//   const [hover, setHover] = React.useState(-1);
+//   const labels = {
+//     0.5: 'Useless',
+//     1: 'Useless+',
+//     1.5: 'Poor',
+//     2: 'Poor+',
+//     2.5: 'Ok',
+//     3: 'Ok+',
+//     3.5: 'Good',
+//     4: 'Good+',
+//     4.5: 'Excellent',
+//     5: 'Excellent+',
+//   }
+//   // const = useState({ rating: $rating })
+//   const [updateActivity, { error }] = useMutation(UPDATE_ACTIVITY);
+//   const itineraryId = localStorage.getItem("itineraryId");
 
-              //or just call update itinearay using state
-              //updateItinerary()
-            }}
-          />
-          {/* <button className="btn rating-submit left">
-              Save rating
-            </button> */}
-        </Box>
-      </form>
-    </div>
-  );
-}
-export default CustomizedRatings;
+//   const handleFormSubmit = async event => {
+//     event.preventDefault();
+//     try {
+//       await updateActivity({
+//         variables: {
+//           rating: props.rating,
+//         }
+//       })
+//     }
+//     catch (e) {
+//       console.log(e);
+//     }
+//   };
+
+
+//   return (
+//     <div>
+//       <form onSubmit={handleFormSubmit}>
+
+//         <Box component="fieldset" mb={3} borderColor="transparent">
+//           <Typography component="legend">How do you like this activity?</Typography>
+//           <Rating
+//             name="simple-controlled"
+//             value={value}
+//             onChange={(event, newValue) => {
+//               setValue(newValue);
+              
+//             }}
+//           />
+         
+//         </Box>
+//       </form>
+//     </div>
+//   );
+// }
+// export default CustomizedRatings;
