--- client/src/pages/SinglePost.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import { useParams } from 'react-router-dom';
-import Createpost from '../components/Ceatepost';
-
-const SinglePost = () => {
-  return (
-    <h1>New itinerary</h1>
-
-  );
-};
-
-export default SinglePost;
